requirements.txt
yfinance==0.2.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
schedule==1.2.1
nsetools==1.0.11
flask==3.0.0
flask-cors==4.0.0